use prettier & build

+++ contracts/test/ERC20.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.15;
 import { SimswapERC20 } from '../SimswapERC20.sol';
 
 contract ERC20 is SimswapERC20 {
-    constructor(uint _totalSupply) {
+    constructor(uint256 _totalSupply) {
         _mint(msg.sender, _totalSupply);
     }
 }
